# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HR Attendance & Meeting Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: month
          in: query
          schema:
            type: ["string", "null"]
          description: "YYYY-MM format to filter by month"
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create an attendance record (check-in/check-out)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Get attendance summary stats
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"

  # ── Meetings ────────────────────────────────────────────────────────────────
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Applications ────────────────────────────────────────────────────────────
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all applications
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully

  # ── Dashboard ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, name, email, department, role, status, joinDate, createdAt]
      properties:
        id:        { type: integer }
        name:      { type: string }
        email:     { type: string }
        phone:     { type: ["string", "null"] }
        department: { type: string }
        position:  { type: ["string", "null"] }
        role:      { type: string, enum: [ADMIN, EMPLOYEE] }
        status:    { type: string, enum: [ACTIVE, INACTIVE, ON_LEAVE] }
        avatarUrl: { type: ["string", "null"] }
        joinDate:  { type: string }
        salary:    { type: ["number", "null"] }
        createdAt: { type: string }
...
    AttendanceRecord:
      type: object
      required: [id, employeeId, date, status, createdAt]
      properties:
        id:           { type: integer }
        employeeId:   { type: integer }
        date:         { type: string }
        checkIn:      { type: ["string", "null"] }
        checkOut:     { type: ["string", "null"] }
        status:       { type: string, enum: [PRESENT, ABSENT, LATE, HALF_DAY, ON_LEAVE] }
        hoursWorked:  { type: ["number", "null"] }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
        employee:
          $ref: "#/components/schemas/Employee"
    AttendanceInput:
      type: object
      required: [employeeId, date, status]
      properties:
        employeeId:  { type: integer }
        date:        { type: string }
        checkIn:     { type: string }
        checkOut:    { type: string }
        status:      { type: string, enum: [PRESENT, ABSENT, LATE, HALF_DAY, ON_LEAVE] }
        hoursWorked: { type: number }
        notes:       { type: string }
    AttendanceUpdate:
      type: object
      properties:
        checkIn:     { type: string }
        checkOut:    { type: string }
        status:      { type: string, enum: [PRESENT, ABSENT, LATE, HALF_DAY, ON_LEAVE] }
        hoursWorked: { type: number }
        notes:       { type: string }
    AttendanceSummary:
      type: object
      required: [totalDays, presentDays, absentDays, lateDays, attendanceRate]
      properties:
        totalDays:      { type: integer }
        presentDays:    { type: integer }
        absentDays:     { type: integer }
        lateDays:       { type: integer }
        halfDays:       { type: integer }
        leaveDays:      { type: integer }
        attendanceRate: { type: number }
...
    Announcement:
      type: object
      required: [id, title, content, author, priority, createdAt]
      properties:
        id:        { type: integer }
        title:     { type: string }
        content:   { type: string }
        author:    { type: string }
        priority:  { type: string, enum: [LOW, NORMAL, HIGH, URGENT] }
        pinned:    { type: boolean }
        createdAt: { type: string }
    AnnouncementInput:
      type: object
      required: [title, content, author]
      properties:
        title:    { type: string, minLength: 1 }
        content:  { type: string, minLength: 1 }
        author:   { type: string }
        priority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT] }
        pinned:   { type: boolean }

    # ── Dashboard ─────────────────────────────────────────────────────────────
...
    Application:
      type: object
      required: [id, organizationId, title, link, status, createdAt]
      properties:
        id:             { type: integer }
        organizationId: { type: integer }
        title:          { type: string }
        link:           { type: string }
        status:         { type: string, enum: [WORKING, FILLED, PENDING] }
        notes:          { type: ["string", "null"] }
        createdAt:      { type: string }
    ApplicationInput:
      type: object
      required: [title, link]
      properties:
        title:  { type: string, minLength: 1 }
        link:   { type: string, minLength: 1 }
        status: { type: string, enum: [WORKING, FILLED, PENDING] }
        notes:  { type: string }
    ApplicationUpdate:
      type: object
      properties:
        title:  { type: string }
        link:   { type: string }
        status: { type: string, enum: [WORKING, FILLED, PENDING] }
        notes:  { type: ["string", "null"] }